moved preallocated to it's own file.
